Actually working example
Previously the example app did not Update the bandit. This is because updating
the bandit in memory means that several instances have to synchronize their
state, and that is probably not how this will be implemented in the end.

However in the interest in having everything working at all times, the example
and http handlers now update bandits in memory.